create new debug module
